cleanup PR, moved non-platform-specific placers
